requirements.txt
# Core data science libraries
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
plotly>=5.3.0
dash>=2.0.0
joblib>=1.1.0

# API and web interaction
requests>=2.26.0
tweepy>=4.10.0
google-api-python-client>=2.47.0
youtube-transcript-api>=0.4.4

# Database
sqlalchemy>=1.4.0

# NLP tools
nltk>=3.6.0

# Utilities
schedule>=1.1.0
python-dotenv>=0.19.0

# Optional monitoring
prometheus-client>=0.14.0